Temporarily place login button in Dashboard
