+++ apps/i15-1/src/routes/Dashboard.tsx
@@ -1,8 +1,14 @@
-import { Container, Typography, Button, Stack } from "@mui/material";
+import { Container, Typography, Stack } from "@mui/material";
 
 import InstrumentSessionView from "../components/InstrumentSessionSelection/InstrumentSessionView.tsx";
+import { useUserAuth } from "../context/userAuth/useUserAuth.ts";
+import { User } from "@diamondlightsource/sci-react-ui";
 
 function Dashboard() {
+  const user = useUserAuth();
+
+  const handleLogIn = () => window.location.assign("/oauth2/sign_in");
+  const handleLogOut = () => window.location.assign("/oauth2/sign_out");
   return (
     <>
       <Container maxWidth="sm" sx={{ mt: 5, mb: 4 }}>
@@ -10,6 +16,16 @@ function Dashboard() {
           <Typography variant="h4" component="h1" textAlign={"center"}>
             Welcome to I15-1
           </Typography>
+          <User
+            onLogin={handleLogIn}
+            onLogout={handleLogOut}
+            user={
+              user.person == null || user.person == undefined
+                ? null
+                : { fedid: user.person }
+            }
+            colour="white"
+          />
           <InstrumentSessionView
             sessionsList={[
               "cm12345-1",
